Update the docker compose configuration
This change revises the Docker Compose configuration, adding in a new
configuration file for the development environment, and moving relevant
configuration settings there from the core configuration file
(docker-compose.yml). It adds in the database dump file, an NGINX
configuration file, and more to bring the Docker Compose more up to date
to reflect the changes in the book.