Rework the Docker configuration
The configuration contained a lot of outdated files which needed to be
updated, unused files which which needed to be removed. This change
handles both of these issues. It also simplifies the configuration by
using Apache with PHP in one container instead of PHP and NGINX in two
separate containers.